* Fixed minigrid tests * Added cmake to dependencies * Added __pycache__ to .gitignore

+++ minigrid/mini_grid.py
@@ -4,6 +4,7 @@ import argparse
 from collections import deque
 from copy import deepcopy
 import logging
+import os
 from pathlib import Path
 import random
 import sys
@@ -324,7 +325,7 @@ def main(args: argparse.Namespace):
     # Copy the domain file to the results folder
     domain_src = Path(__file__).parent / 'domain.pddl'
     domain_dst = args.results / 'domain.pddl'
-    if domain_src != domain_dst:
+    if domain_src != domain_dst and not os.path.exists(domain_dst):
         import shutil
         shutil.copy(domain_src, domain_dst)
         logger.info(f'Copied domain file to {domain_dst}')
